+++ packages/zep-script-cli/src/commands/index.ts
@@ -2,6 +2,7 @@ import archive from "./archive";
 import build from "./build";
 import publish from "./publish";
 import init from "./init";
+import pack from "./pack";
 
 export type OptionValue = string | boolean | number;
 
@@ -24,4 +25,4 @@ export type Command = {
   func: CommandFunction<Object>;
 };
 
-export const commands = [init, build, publish, archive] as Command[];
+export const commands = [init, build, publish, archive, pack] as Command[];